requirements.txt
safetensors==0.7.0
torch>=2.10.0
torchvision>=0.25.0
numpy
opencv_python==4.13.0.92
Pillow==12.1.1
accelerate==1.12
controlnet_aux==0.0.10
diffusers==0.36.0
peft>=0.18.1
transformers>=5.1.0
einops>=0.8.2
click>=8.3.1
ipycanvas>=0.14.3
ipython>=8.12.3
ipywidgets>=8.1.8
